Change all system clock types to int64_t in bitrate_controller.
They are both compared to int64_t types inside the class, and is being called
with int64_t types. Could possibly cause bugs.

R=pbos@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/33529004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@7832 4adac7df-926f-26a2-2b94-8c16560cd09d

+++ webrtc/modules/bitrate_controller/send_side_bandwidth_estimation.cc
@@ -103,7 +103,7 @@ void SendSideBandwidthEstimation::UpdateReceiverEstimate(uint32_t bandwidth) {
 void SendSideBandwidthEstimation::UpdateReceiverBlock(uint8_t fraction_loss,
                                                       uint32_t rtt,
                                                       int number_of_packets,
-                                                      uint32_t now_ms) {
+                                                      int64_t now_ms) {
   if (first_report_time_ms_ == -1)
     first_report_time_ms_ = now_ms;
 
@@ -163,7 +163,7 @@ void SendSideBandwidthEstimation::UpdateUmaStats(int64_t now_ms,
   }
 }
 
-void SendSideBandwidthEstimation::UpdateEstimate(uint32_t now_ms) {
+void SendSideBandwidthEstimation::UpdateEstimate(int64_t now_ms) {
   // We trust the REMB during the first 2 seconds if we haven't had any
   // packet loss reported, to allow startup bitrate probing.
   if (ProbingExperimentIsEnabled()) {
@@ -230,7 +230,7 @@ bool SendSideBandwidthEstimation::IsInStartPhase(int64_t now_ms) const {
          now_ms - first_report_time_ms_ < kStartPhaseMs;
 }
 
-void SendSideBandwidthEstimation::UpdateMinHistory(uint32_t now_ms) {
+void SendSideBandwidthEstimation::UpdateMinHistory(int64_t now_ms) {
   // Remove old data points from history.
   // Since history precision is in ms, add one so it is able to increase
   // bitrate if it is off by as little as 0.5ms.
